# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Pramana API
  version: 0.1.0
  summary: Compliance training and tracking platform
  description: |
    The Pramana v1 API. Single-tenant deployment for John Thomas Corporate,
    scoped to **SOX (Sarbanes-Oxley)** compliance training.

    Authentication is via OIDC SSO; the bearer token is a JWT issued by the
    customer's IdP. All endpoints require authentication unless noted as
    `[Public]`.

    The **content pipeline** — commission a Package Request → Mentible
    manufactures a Consumable Package → Pramana ingests, a human approves, and
    it publishes to an immutable course version — is defined by
    `docs/03_ai_drafted_human_approved_content.md` and Mentible **ADR-011**.

    See `docs/02_resolved_decisions.md` for the canonical specification this
    API implements.
  contact:
    name: WeGoFwd
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.pramana.example.com/v1
    description: Production
  - url: https://staging-api.pramana.example.com/v1
    description: Staging
  - url: http://localhost:8000/v1
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /content-drafts:
    get:
      tags: [ContentDrafts]
      summary: List content drafts (the review queue)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ContentDraftStatus"
        - name: framework
          in: query
          schema:
            type: string
        - name: quarantined
          in: query
          description: When true, return only packages that failed verification.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Page of drafts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDraftPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /content-drafts/{draft_id}:
    parameters:
      - $ref: "#/components/parameters/DraftId"
    get:
      tags: [ContentDrafts]
      summary: Get a draft for review (preview payload)
      description: |
        Returns the full review payload: modules + quiz + per-claim citations,
        provenance, the signature/content_hash verification result, and artifact
        references. Used by the review screen (US-PLATFORM-0004).
      responses:
        "200":
          description: Draft detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDraftDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /content-drafts/{draft_id}/submit-for-review:
    parameters:
      - $ref: "#/components/parameters/DraftId"
    post:
      tags: [ContentDrafts]
      summary: Open a draft for review
      description: Transition `RECEIVED`/`DRAFT` → `IN_REVIEW`.
      responses:
        "200":
          description: Draft in review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDraft"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Not in a pre-review state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /content-drafts/{draft_id}/approve:
    parameters:
      - $ref: "#/components/parameters/DraftId"
    post:
      tags: [ContentDrafts]
      summary: Approve a draft (attestation + separation of duties)
      description: |
        Transition `IN_REVIEW` → `APPROVED`. The approver **must not** be the
        draft's requester/generator (separation of duties → 409). Captures an
        attestation and freezes the content hash. See US-FCPA-0005 / US-PLATFORM-0004.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRequest"
      responses:
        "200":
          description: Draft approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDraft"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Not in review, or separation-of-duties violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /content-drafts/{draft_id}/request-changes:
    parameters:
      - $ref: "#/components/parameters/DraftId"
    post:
      tags: [ContentDrafts]
      summary: Request changes on a draft
      description: Transition `IN_REVIEW` → `DRAFT`/`RECEIVED` with reviewer notes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewNotes"
      responses:
        "200":
          description: Changes requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDraft"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Not in review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /content-drafts/{draft_id}/reject:
    parameters:
      - $ref: "#/components/parameters/DraftId"
    post:
      tags: [ContentDrafts]
      summary: Reject a draft (terminal)
      description: Transition `IN_REVIEW` → `REJECTED`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewNotes"
      responses:
        "200":
          description: Draft rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDraft"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Not in review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /content-drafts/{draft_id}/publish:
    parameters:
      - $ref: "#/components/parameters/DraftId"
    post:
      tags: [ContentDrafts]
      summary: Publish an approved draft to an immutable course version
      description: |
        Transition `APPROVED` → `PUBLISHED`, materializing a `CourseVersion` that
        becomes assignable. Returns the created version.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequest"
      responses:
        "201":
          description: Published course version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseVersion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Not in APPROVED state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /content-drafts/{draft_id}/regenerate:
    parameters:
      - $ref: "#/components/parameters/DraftId"
    post:
      tags: [ContentDrafts]
      summary: Regenerate a draft with updated parameters
      description: |
        Re-issues the original Package Request (optionally with parameter overrides)
        to Mentible; the returned package arrives as a **new `package_version`** that
        supersedes this draft. Does not rewrite published versions or completion
        records. Lineage handling tracks ADR-011 §10. See US-PLATFORM-0005.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateRequest"
      responses:
        "202":
          description: Regeneration request pushed to Mentible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentRequest"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Player (Present phase — playback + watch-progress)
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC-issued JWT in the `Authorization: Bearer <token>` header. The
        token's `sub` claim is mapped to a Pramana `user_id` at first login.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    DraftId:
      name: draft_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not authorised
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Input failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code
          example: cooldown_active
        message:
          type: string
          example: Cooldown is active until 2027-01-15.
        context:
          type: object
          additionalProperties: true
    ValidationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                  message:
                    type: string
    Pagination:
      type: object
      required: [page, page_size, total]
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
...
    CourseVersion:
      type: object
      required:
        [
          version_id,
          course_id,
          version_number,
          published_at,
          is_active,
        ]
      properties:
        version_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        version_number:
          type: integer
          minimum: 1
        published_at:
          type: string
          format: date-time
        published_by_user_id:
          type: string
          format: uuid
        video_asset_id:
          type: string
        min_watch_pct:
          type: integer
          minimum: 0
          maximum: 100
          default: 0
        is_active:
          type: boolean
        is_material_change:
          type: boolean
          default: false
...
    SourceDefinition:
      type: object
      required: [framework, clause]
      properties:
        framework:
          type: string
          example: fcpa
        clause:
          type: string
          example: anti-bribery
        ref:
          type: string
          example: docs/frameworks/framework_fcpa.md#anti-bribery
...
    ContentRequestStatus:
      type: string
      enum: [requested, generating, received, in_review, published, failed]
...
    ContentRequest:
      type: object
      required: [request_id, framework, title, status, requested_by, created_at]
      properties:
        request_id:
          type: string
          format: uuid
        framework:
          type: string
        title:
          type: string
        status:
          $ref: "#/components/schemas/ContentRequestStatus"
        requested_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        package_id:
          type: string
          format: uuid
          nullable: true
          description: Set once Mentible returns a package.
        draft_id:
          type: string
          format: uuid
          nullable: true
          description: The ContentDraft created on ingestion.
        satisfies_stories:
          type: array
          items:
            type: string
...
    Provenance:
      type: object
      required: [engine, model, provider, prompt_version, generated_at]
      properties:
        engine:
          type: string
          example: mentible
        model:
          type: string
          example: claude-sonnet-4-6
        provider:
          type: string
          example: anthropic
        prompt_version:
          type: string
          example: psai-2026-06
        generated_at:
          type: string
          format: date-time
...
    ContentDraftStatus:
      type: string
      enum: [received, draft, in_review, approved, published, rejected]
    VerificationResult:
      type: object
      required: [signature_ok, content_hash_ok]
      properties:
        signature_ok:
          type: boolean
        content_hash_ok:
          type: boolean
        quarantined:
          type: boolean
          description: True if verification failed; the draft is not publishable.
    ContentDraft:
      type: object
      required: [draft_id, status, title, framework]
      properties:
        draft_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: "#/components/schemas/ContentDraftStatus"
        title:
          type: string
        framework:
          type: string
        package_id:
          type: string
          format: uuid
          nullable: true
        package_version:
          type: integer
          nullable: true
        verified:
          type: boolean
          description: Signature + content_hash verified on ingest.
        created_at:
          type: string
          format: date-time
    ContentDraftDetail:
      allOf:
        - $ref: "#/components/schemas/ContentDraft"
        - type: object
          properties:
            provenance:
              $ref: "#/components/schemas/Provenance"
            source_definitions:
              type: array
              items:
                $ref: "#/components/schemas/SourceDefinition"
            modules:
              type: array
              items:
                type: object
                additionalProperties: true
            quiz:
              type: object
              additionalProperties: true
            artifacts:
              type: array
              items:
                type: object
                additionalProperties: true
            assets:
              type: array
              items:
                type: object
                additionalProperties: true
            verification:
              $ref: "#/components/schemas/VerificationResult"
            review_notes:
              type: string
              nullable: true
    ContentDraftPage:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContentDraft"
        pagination:
          $ref: "#/components/schemas/Pagination"
    ApproveRequest:
      type: object
      required: [attestation_text]
      properties:
        attestation_text:
          type: string
          description: The approver's accuracy attestation, captured as audit evidence.
    ReviewNotes:
      type: object
      required: [notes]
      properties:
        notes:
          type: string
    PublishRequest:
      type: object
      properties:
        effective_from:
          type: string
          format: date-time
          nullable: true
        is_material_change:
          type: boolean
          default: false
    RegenerateRequest:
      type: object
      properties:
        notes:
          type: string
        parameter_overrides:
          type: object
          description: |
            Partial Package Request fields to change before regenerating (e.g.
            scope, assessment, constraints). Unspecified fields reuse the original.
          additionalProperties: true

    # ----- Player (playback + watch progress) --------------------------
...